Programmation_Application_Distribuees_Reseaux/Projet/Code/requirements.txt
cffi==1.15.1
cryptography==41.0.3
pycparser==2.21
pyOpenSSL==23.2.0
